Add ascii to horse code conversion
New controller with a field which is bound to the editable input
element and a function for ascii to horse code conversion which
produces the output of the read only input element.

Uses a function with a static switch statement to return the correct
horse emoji sequence for a character and returns unicorn for unknown
characters. Case is not preserved.

Some super lazy css from memory (worked first time!) more interesting
css to follow.