Add decoding of horse code to ascii, some better styling
Added a CSS class for centering stuff on the page, used it to replace
the global style setting for html elements.

Modified the encoding of horse-code messages so that they would be
decodable. Then produced the correct decoding function and a button to
swap between which encoding function is currently in use.