Issue #2 Inputs are not debounced
Debounce the onChange event for the text input so dispatch is only called after
500ms of no inputs.

Use defaultValue instead of value and let the browser deal with managing the
contents of the TextField. I think this optimisation is actually contributing
more than the debouncing but the debouncing certainly does no harm. Add a key
to the TextField so that the toggle forces a re-render.