Restore debounced change handling
Rather than manipulate the text value directly in the reducer which
causes a lot of re-renders as we re-create the textareas, append to the
value of the underlying textarea directly and let the usual event
handling sort it out. This means appending using the horse buttons is
effectively the same as pressing the key on the keyboard so the
translation will fire at most 1/500ms.